requirements.txt
pandas==1.2.4
requests==2.22.0
beautifulsoup4==4.9.3
Flask==1.1.2
flask_sqlalchemy==2.5.1
gspread==3.7.0
mysql_connector_repackaged==0.3.1
oauth2client==4.1.3
